requirements.txt
python-dotenv
streamlit
watchdog
pydantic
azure-storage-blob
azure-identity
